Migrate remaining version related functions
